requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdf2image==1.17.0
minio==7.2.3
python-dotenv==1.0.0
Pillow>=10.0.0
